+++ packages/aws-amplify-angular/src/components/authenticator/forgot-password-component/forgot-password.class.ts
@@ -1,0 +1,3 @@
+export class ForgotPasswordClass {
+  data: any;
+}