+++ packages/aws-amplify/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.4.9-unstable.1"></a>
+## [0.4.9-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify@0.4.9-unstable.0...aws-amplify@0.4.9-unstable.1) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="0.4.9-unstable.0"></a>
+## [0.4.9-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify@0.4.8...aws-amplify@0.4.9-unstable.0) (2018-07-02)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="0.4.8"></a>
+## [0.4.8](https://github.com/aws/aws-amplify/compare/aws-amplify@0.4.8-unstable.5...aws-amplify@0.4.8) (2018-06-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="0.4.8-unstable.5"></a>
+## [0.4.8-unstable.5](https://github.com/aws/aws-amplify/compare/aws-amplify@0.4.8-unstable.4...aws-amplify@0.4.8-unstable.5) (2018-06-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
 <a name="0.4.8-unstable.4"></a>
 ## [0.4.8-unstable.4](https://github.com/aws/aws-amplify/compare/aws-amplify@0.4.8-unstable.3...aws-amplify@0.4.8-unstable.4) (2018-06-29)